backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
sentence-transformers==3.0.0
faiss-cpu==1.9.0.post1
requests==2.32.0
numpy==1.26.4
python-dotenv==1.0.1
python-multipart==0.0.12
pypdf==4.3.0
langchain-community==0.3.7
unstructured==0.16.3
pytesseract==0.3.13
pdf2image==1.17.0
Pillow>=10.0.0
python-docx==1.1.2
PyJWT>=2.9.0
pytest==8.3.0
httpx==0.27.0
